Create webpack-notes.mdx
I don't think this is actually integrated into the docs, but we'll see (sorry, done on GitHub). Also, I'm aware that the text here is really uneven. It probably goes into too much detail about the module federation setup and could be organised better, especially as the code-snippet at the end should probably be surfaced better.